Move the rest frontend end image files to `.~/images`.
