+++ js/src/product-feed/submission-success-guide/index.js
@@ -21,7 +21,7 @@ import AddPaidCampaignButton from '.~/components/paid-ads/add-paid-campaign-butt
 import { glaData, GUIDE_NAMES, LOCAL_STORAGE_KEYS } from '.~/constants';
 import localStorage from '.~/utils/localStorage';
 import { getProductFeedUrl } from '.~/utils/urls';
-import wooLogoURL from './woocommerce-logo.svg';
+import wooLogoURL from '.~/images/logo/woocommerce-logo.svg';
 import googleLogoURL from '.~/images/google-logo.svg';
 import { recordGlaEvent } from '.~/utils/tracks';
 import './index.scss';
